Readme.md: update build parallellism information
BaseTools can now be built in parallel, so drop the statement that says
it cannot, and add -j flag to example command line.

Also, the `build` command now utilizes parallellism by default, so update
text and examples to reflect this (and drop the examples for manually
figuring out available core count).

Signed-off-by: Leif Lindholm <leif@nuviainc.com>